+++ CHANGES.md
@@ -1,5 +1,6 @@
-Change log
+Changes
 ==========
 
+* 09-01-2016: Added `resubscribe` endpoint
 * 03-11-2015: Remove development_mode
 * 20-12-2014: Fix for track function when no data are being tracked